Update the project availability
Hello @k3b, I know you are not the upstream, but your fork contains the latest commits I found. Do you know if this project is suppose to work in LineageOS 14.1 (7.1.2)? It does not work for me and the only info I found is https://stackoverflow.com/questions/45163799/com-android-camera-new-picture-is-deprecated-for-all-apps-android-n

Do you know of a replacement or somebody who could help to resurrect this awesome project? Paid version is fine by me (decent amount).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>